requirements.txt
Flask==3.0.0
python-barcode==0.15.1
Pillow==9.5.0
opencv-python==4.8.1.78
pyzbar==0.1.9
